+++ internal/adapter/platform/linux/libei_client.h
@@ -28,8 +28,10 @@ int neru_ei_move_abs(NeruEiClient *c, int x, int y);
 int neru_ei_button(NeruEiClient *c, int button, int pressed);
 
 // Emit a scroll event. axis: 0 = vertical, 1 = horizontal. delta is the scroll
-// distance in logical pixels (positive = down/right). Returns 1 on success.
-int neru_ei_scroll(NeruEiClient *c, int axis, int delta);
+// distance in logical pixels (positive = down/right), and is a double because
+// ei_device_scroll_delta is pixel-precise: a fraction of a wheel notch is a
+// legal delta, which is what an animated scroll sends. Returns 1 on success.
+int neru_ei_scroll(NeruEiClient *c, int axis, double delta);
 
 // Press or release a keyboard key (evdev keycode). Returns 1 on success, 0 when
 // no keyboard device is available on the granted session.
